merge

+++ hr-app/src/main/java/com/remark_herlan/hr_app/controller/RolesController.java
@@ -7,7 +7,6 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestAttribute;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
@@ -37,7 +36,7 @@ public class RolesController {
 	}
 
 	@GetMapping("/{id}")
-	public ResponseInfo<Optional<Roles>> getMethod(@PathVariable Integer id)
+	public ResponseInfo<Optional<Roles>> getMethod(@PathVariable Long id)
 			throws InternalServerException, DataNotFoundException {
 		return service.getInfo(id);
 	}
